refactor: access, refresh 분리
